MACRO: reuse PSI factory
