+++ src/main/kotlin/org/rust/lang/core/psi/ext/PsiElement.kt
@@ -51,6 +51,10 @@ inline fun <reified T : PsiElement> PsiElement.contextOrSelf(): T? =
     PsiTreeUtil.getContextOfType(this, T::class.java, /* strict */ false)
 
 
+inline fun <reified T : PsiElement> PsiElement.childrenOfType(): List<T> =
+    PsiTreeUtil.getChildrenOfTypeAsList(this, T::class.java)
+
+
 inline fun <reified T : PsiElement> PsiElement.descendantOfTypeStrict(): T? =
     PsiTreeUtil.findChildOfType(this, T::class.java, /* strict */ true)
